REFACTOR : HttpResponse 클래스로 응답 관련 코드 이동
