+++ src/main/java/http/HttpResponse.java
@@ -1,5 +1,42 @@
 package http;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import webserver.RequestHandler;
+
+import java.io.DataOutputStream;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+
 public class HttpResponse {
-    //TODO
+    private static final Logger logger = LoggerFactory.getLogger(HttpResponse.class);
+    private static final String requestFilePath = "./src/main/resources/templates/";
+    private String uri;
+
+    public void send(DataOutputStream dos) throws IOException {
+        byte[] body = Files.readAllBytes(new File(requestFilePath + uri).toPath());
+        response200Header(dos, body.length);
+        responseBody(dos, body);
+    }
+
+    private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
+        try {
+            dos.writeBytes("HTTP/1.1 200 OK \r\n");
+            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
+            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
+            dos.writeBytes("\r\n");
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
+    }
+
+    private void responseBody(DataOutputStream dos, byte[] body) {
+        try {
+            dos.write(body, 0, body.length);
+            dos.flush();
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
+    }
 }
