+++ src/main/java/webserver/RequestHandler.java
@@ -13,7 +13,6 @@ import java.nio.file.Files;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
-    private static final String requestFilePath = "./src/main/resources/templates/";
 
     private Socket connection;
 
@@ -28,34 +27,14 @@ public class RequestHandler implements Runnable {
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             // TODO 사용자 요청에 대한 처리는 이 곳에 구현하면 된다.
             HttpRequest httpRequest = new HttpRequest(new BufferedReader(new InputStreamReader(in, StandardCharsets.UTF_8)));
-            DataOutputStream dos = new DataOutputStream(out);
+
             HttpResponse httpResponse = new HttpResponse();
             Controller controller = ControllerHandler.handleController(httpRequest);
             controller.makeResponse(httpRequest, httpResponse);
 
-            byte[] body = Files.readAllBytes(new File(requestFilePath + httpRequest.getUri()).toPath());
-            response200Header(dos, body.length);
-            responseBody(dos, body);
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
+            DataOutputStream dos = new DataOutputStream(out);
+            httpResponse.send(dos);
 
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
-    }
-
-    private void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
         } catch (IOException e) {
             logger.error(e.getMessage());
         }
